safari-600.7-branch/LayoutTests/svg/animations/svglength-animation-px-to-cm-expected.txt

SVG 1.1 dynamic animation tests

Test SVGLength animation from px to cm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 144.5
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 189
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
